invoice number should be unique
We are having problem deleting invoice with duplicated number.

Because we are using invoice number in the url, when there are 2 invoices with the same number, random invoice are retrieved and/or deleted. We were deciding weather to use _id in the URL or restrict the invoiceNumber to be unique. We decided to go with unique invoice number to keep the pretty URL. Which means when we delete an invoice, we need to update invoice number and mark it as deleted.